LayoutTests/media/modern-media-controls/slider/slider-styles-expected.txt

Testing the Slider computed style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS containerStyle.position is "absolute"
PASS canvasStyle.position is "absolute"
PASS canvasStyle.top is "0px"
PASS canvasStyle.left is "0px"
PASS canvasStyle.width is "200px"
PASS canvasStyle.height is "25px"
PASS canvasStyle.pointerEvents is "none"
PASS inputStyle.position is "absolute"
PASS inputStyle.top is "0px"
PASS inputStyle.left is "0px"
PASS inputStyle.width is "200px"
PASS inputStyle.height is "25px"
PASS inputStyle.margin is "0px"
PASS inputStyle.webkitAppearance is "none"
PASS inputStyle.backgroundColor is "rgba(0, 0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
